cs: object indexes, wip

+++ json-sick-csharp/SickSharp/Format/VarTable.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Diagnostics;
 using System.IO;
 using SickSharp.Primitives;
@@ -6,21 +7,21 @@ namespace SickSharp.Format
 {
     public abstract class BasicVarTable<TV>
     {
-        private readonly int _dataOffset;
-        private readonly int _sizeOffset;
+        private readonly UInt32 _dataOffset;
+        private readonly UInt32 _sizeOffset;
         protected readonly Stream Stream;
 
 
-        public BasicVarTable(Stream stream, int offset)
+        public BasicVarTable(Stream stream, UInt32 offset)
         {
             Stream = stream;
-            Count = Stream.ReadInt32(offset);
+            Count = (uint)Stream.ReadInt32(offset);
 
             _sizeOffset = offset + sizeof(int);
             _dataOffset = _sizeOffset + sizeof(int) * (Count + 1);
         }
 
-        public int Count { get; }
+        public UInt32 Count { get; }
 
 
         public TV Read(int index)
@@ -28,26 +29,26 @@ namespace SickSharp.Format
             Debug.Assert(index < Count);
             var szTarget = _sizeOffset + index * sizeof(int);
             var szTargetNext = _sizeOffset + (index + 1) * sizeof(int);
-            var relativeDataOffset = Stream.ReadInt32(szTarget);
-            var endOffset = Stream.ReadInt32(szTargetNext);
+            var relativeDataOffset = (UInt32)Stream.ReadInt32(szTarget);
+            var endOffset = (UInt32)Stream.ReadInt32(szTargetNext);
 
             var absoluteStartOffset = _dataOffset + relativeDataOffset;
             var byteLen = endOffset - relativeDataOffset;
             return BasicRead(absoluteStartOffset, byteLen);
         }
 
-        protected abstract TV BasicRead(int absoluteStartOffset, int byteLen);
+        protected abstract TV BasicRead(UInt32 absoluteStartOffset, UInt32 byteLen);
     }
 
     public abstract class VarTable<TV> : BasicVarTable<TV>
     {
-        protected VarTable(Stream stream, int offset) : base(stream, offset)
+        protected VarTable(Stream stream, UInt32 offset) : base(stream, offset)
         {
         }
 
-        protected override TV BasicRead(int absoluteStartOffset, int byteLen)
+        protected override TV BasicRead(UInt32 absoluteStartOffset, UInt32 byteLen)
         {
-            var bytes = Stream.ReadBuffer(absoluteStartOffset, byteLen);
+            var bytes = Stream.ReadBuffer(absoluteStartOffset, (int)byteLen);
             return Convert(bytes);
         }
